refactor: 테스트코드에서만 사용되는 메서드 제거
